# Backend/swagger-online/swagger.yaml
swagger: '2.0'
info:
  title: Payment Routes
  description: Apis for Payment Related Service
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
host: localhost:4000
basePath: /online
schemes:
  - http
  - https
paths:
  /health:
    get:
      summary: check health of server
      responses:
        200:
          description: OK
        default:  
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /user/register:
    post:
      summary: api to create a user -> merchant/customer
      parameters:
        - $ref: '#/parameters/merchantXApiKey'
        - $ref: '#/parameters/appId'
        - $ref: '#/parameters/appSecret'
        - $ref: '#/parameters/registerBody'
      produces:
        - application/json
      responses:
        200:
          description: OK
        default:  
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /user/login:
    post:
      summary: api to create a user -> merchant/customer
      parameters:
        - $ref: '#/parameters/loginBody'
      produces:
        - application/json
      responses:
        200:
          description: OK
        default:  
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /user/verifyOTP:
    post:
      summary: api to verify OTP
      parameters:
        - $ref: '#/parameters/otpBody'
      produces:
        - application/json
      responses:
        200:
          description: OK
        default:  
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
            
parameters:
  billId: &billId
    in: path
    name: billId
    type: string
    required: true
  documentStructure: &documentStructure
    name: body
    in: body
    required: true
    schema:
      $ref: '#/definitions/documentStructure'
  merchantXApiKey: &merchantXApiKey
    name: x-api-key
    description: enter merchant x-api-key
    in: header
    required: false
    type: string
  appId: &appId
    name: app-id
    description: enter app-id
    in: header
    required: true
    type: string
  adminSecret: &adminSecret
    name: admin-secret
    description: enter admin-secret base64 encoded <admin_username>:<admin_password>
    in: header
    required: true
    type: string
  appSecret: &appSecret
    name: app-secret
    description: enter app-secret
    in: header
    required: true
    type: string  
  appBody: &appBody
    name: body
    in: body
    required: true
    schema: 
      $ref: '#/definitions/appBody'
  appEditBody: &appEditBody
    name: body
    in: body
    required: true
    schema: 
      $ref: '#/definitions/appEditBody'
  mockPayBody: &mockPayBody
    name: body
    in: body
    required: true
    schema: 
      $ref: '#/definitions/mockPayBody'
  billBody: &billBody
    name: body
    in: body
    required: true
    schema: 
      $ref: '#/definitions/billBody'
  registerBody: &registerBody
    name: body
    in: body
    required: true
    schema:
      $ref: '#/definitions/registerBody'
  loginBody: &loginBody
    name: body
    in: body
    required: true
    schema:
      $ref: '#/definitions/loginBody'
  otpBody: &otpBody
    name: body
    in: body
    required: true
    schema:
      $ref: '#/definitions/otpBody'

definitions: 
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  documentStructure:
    required:
      - document_type
      - document_info
    properties:
      document_type:
        type: string
        enum: 
          - pan
          - aadhar
          - bank_account
          - vpa
      document_info:
        type: object

  billBody:
    required:
     - customerId
     - billerId
     - additionalInfo
     - amountRanges
     - paymentGateway
    properties:
      customerId:
        type: string
      billerId:
        type: string
      additionalInfo:
        type: object
        properties:
          key:
            type: string
      amountRanges:
        type: object
        properties:
          minAmount:
            type: integer
          maxAmount:
            type: integer
      paymentGateway:
        type: string
        enum:
          - setu
          - razorpay
  paymentSetMockPay:
    properties:
      amount_paid:
        type: integer
      created_at:
        type: string
  appBody:
    required:
     - appName
    properties:
      appName:
        type: string
  appEditBody:
    required:
     - callbackUrl
     - callbackSecret
    properties:
      callbackUrl:
        type: string
      callbackSecret:
        type: string
  mockPayBody:
    required:
      - platform_id
      - payment_set
      - first_min_amount
      - payee_name
      - max_amount
      - status
      - biller_id
      - callback_url
      - payment_link
    properties:
      platform_id:
        type: string
      payment_set:
        type: array
        items:
          $ref: '#/definitions/paymentSetMockPay'
      first_min_amount:
        type: integer
      payee_name:
        type: string
      max_amount:
        type: integer
      status: 
        type: string
        enum:
          - partially_paid
          - paid
          - failed
      biller_id: 
        type: string
      callback_url: 
        type: string
      payment_link:
        type: string
  otpBody:
    required:
      - otp
      - phoneNumber
      - hash
    properties:
      otp:
        type: string
      phoneNumber:
        type: string
      hash: 
        type: string
  loginBody:
    required:
      - phoneNumber
      - role
    properties:
      phoneNumber:
        type: string
      role:
        type: string
        enum:
          - merchant
          - customer
      
  registerBody:
    required:
      - phoneNumber
    properties:
      phoneNumber:
        type: string
      displayName:
        type: string  
      role:
        type: string
        enum:
          - merchant
          - customer